Determining the optimal value for the second charge trapping method was taking excessively long due to energy calibration with a very large number of bins.  Reduced the number of bins to make this run faster.
Fixed a bug resulting in the channel number being set to zero for all events after the end of the first run when processing multiple runs at the same time.

Fixed the computation of the DCR slope as a function of energy.

Fixed a typo in the naming of the histograms used for charge trapping which resulted in a seg fault on the calibration step.